Rename pnp classes and relevant files
